clippy fixes:logical expr mod
