+++ brahmand/src/query_planner/analyzer/graph_context.rs
@@ -5,7 +5,7 @@ use crate::{
             analyzer_pass::AnalyzerResult,
             errors::{AnalyzerError, Pass},
         },
-        logical_expr::logical_expr::Direction,
+        logical_expr::Direction,
         logical_plan::GraphRel,
         plan_ctx::{PlanCtx, TableCtx},
     },
