+++ brahmand/src/query_planner/analyzer/projection_tagging.rs
@@ -7,7 +7,7 @@ use crate::{
             analyzer_pass::{AnalyzerPass, AnalyzerResult},
             errors::{AnalyzerError, Pass},
         },
-        logical_expr::logical_expr::{
+        logical_expr::{
             AggregateFnCall, Column, LogicalExpr, PropertyAccess, TableAlias,
         },
         logical_plan::{LogicalPlan, Projection, ProjectionItem},
