+++ brahmand/src/query_planner/logical_plan/where_clause.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::{
     open_cypher_parser::ast::WhereClause,
     query_planner::{
-        logical_expr::logical_expr::LogicalExpr,
+        logical_expr::LogicalExpr,
         logical_plan::{Filter, LogicalPlan},
     },
 };
